refactor(workbooks): decompose WorkbookGrid — extract Filter + Summary panels (EP-GRID-WORKBOOKS)
Grid.tsx had grown to 1379 LOC (over the 1000 hard cap) because every parity
feature added another inline toolbar panel, making it a merge-conflict magnet —
any two in-flight grid PRs collide on it. Start decomposing the panels into typed,
presentational subcomponents so a future panel lands as its own file (an
independent PR) instead of another block in Grid.tsx.

- New GridPanels.tsx: GridFilterPanel (structured filter builder) + GridSummaryPanel
  (group-by pivot-lite), pure UI over typed props the Grid owns.
- Grid.tsx renders <GridFilterPanel/> / <GridSummaryPanel/> and drops the extracted
  imports; 1379 -> 1154 LOC, baseline ratcheted down.
- No behavior change — identical rendering. 106 workbooks unit tests still pass.
- Remaining panels (Format, Group, Columns) follow the same mechanical pattern in
  fast follow-ups.

Backlog: BI-F820069E. Local typecheck skipped (source-only worktree; CI's Typecheck
is the authoritative gate for the prop wiring).

UX-Fit-Decision: pure refactor — the Filter and Summary panels render exactly as
before (same controls, same progressive-disclosure toolbar toggles). No user-facing
change, so human_cognitive_load is unchanged; the extraction only moves JSX into
typed components. Per AGENTS.md §12/§17 this is a no-UX-change structural edit.

Design-Grounding-Decision:
- Existing specs/plans reviewed: docs/superpowers/plans/2026-06-11-universal-grid-workbooks-phase2.md (the "decompose Grid.tsx" debt note in the Remaining section).
- Current code substrate reviewed: apps/web/components/workbooks/Grid.tsx (the inline panels), apps/web/components/workbooks/grid-filter-builder.ts, apps/web/components/workbooks/grid-summary.ts.
- Source of truth: the existing panel JSX + the state the Grid owns; components take it as typed props.
- Decision: mechanical extraction to subcomponents, identical behavior — no contract, routing, or data-model change.


Process-Spine-Decision: pure structural refactor — extract two existing toolbar panels (Filter, Summary) from Grid.tsx into GridPanels.tsx with identical behavior. No new UX/workflow/queue surface; the new module is a presentational split of existing code, so no spec/plan change is warranted (recorded here per the Spec/Plan/Doc gate). Reporting-composition baseline re-generated because the Summary panel's existing (grandfathered) <table> moved files.
