+++ scripts/reporting-composition-baseline.json
@@ -6,20 +6,12 @@
   "apps/web/app/(shell)/admin/scheduled-jobs/ScheduledJobsClient.tsx": 1,
   "apps/web/app/(shell)/coworker-decisions/decisions/[interactionId]/page.tsx": 1,
   "apps/web/app/(shell)/finance/assets/[id]/page.tsx": 1,
-  "apps/web/app/(shell)/finance/assets/page.tsx": 1,
   "apps/web/app/(shell)/finance/banking/[id]/page.tsx": 1,
   "apps/web/app/(shell)/finance/bills/[id]/page.tsx": 2,
-  "apps/web/app/(shell)/finance/bills/page.tsx": 1,
   "apps/web/app/(shell)/finance/expense-claims/[id]/page.tsx": 1,
-  "apps/web/app/(shell)/finance/expense-claims/page.tsx": 1,
   "apps/web/app/(shell)/finance/invoices/[id]/page.tsx": 2,
-  "apps/web/app/(shell)/finance/invoices/page.tsx": 1,
-  "apps/web/app/(shell)/finance/my-expenses/page.tsx": 1,
-  "apps/web/app/(shell)/finance/payment-runs/page.tsx": 1,
   "apps/web/app/(shell)/finance/purchase-orders/[id]/page.tsx": 2,
-  "apps/web/app/(shell)/finance/purchase-orders/page.tsx": 1,
   "apps/web/app/(shell)/finance/recurring/[id]/page.tsx": 2,
-  "apps/web/app/(shell)/finance/recurring/page.tsx": 1,
   "apps/web/app/(shell)/finance/reports/aged-creditors/page.tsx": 1,
   "apps/web/app/(shell)/finance/reports/aged-debtors/page.tsx": 1,
   "apps/web/app/(shell)/finance/reports/cash-flow/page.tsx": 1,
@@ -30,7 +22,6 @@
   "apps/web/app/(shell)/finance/reports/vat-summary/page.tsx": 1,
   "apps/web/app/(shell)/finance/settings/currency/page.tsx": 1,
   "apps/web/app/(shell)/finance/suppliers/[id]/page.tsx": 2,
-  "apps/web/app/(shell)/finance/suppliers/page.tsx": 1,
   "apps/web/app/(shell)/ops/dev-loop/page.tsx": 2,
   "apps/web/app/(shell)/ops/security/page.tsx": 1,
   "apps/web/app/(shell)/platform/ai/runtime-health/page.tsx": 1,
@@ -80,6 +71,6 @@
   "apps/web/components/product/ProductSupplyChainPanel.tsx": 1,
   "apps/web/components/storefront/sections/LoanCalculatorSection.tsx": 1,
   "apps/web/components/wiki/WikiPageViewer.tsx": 1,
-  "apps/web/components/workbooks/Grid.tsx": 1,
+  "apps/web/components/workbooks/GridPanels.tsx": 1,
   "apps/web/components/workspace/PlatformReadinessMatrix.tsx": 1
 }
